[MISC] *: move mail features to viin_brand_mail
