--- viin_brand/models/__init__.py
@@ -1,2 +1,0 @@
-from . import mail_template
-from . import ir_ui_view
