+++ viin_brand_mail/models/__init__.py
@@ -1,1 +1,3 @@
 from . import mail_alias
+from . import mail_template
+from . import ir_ui_view
