+++ viin_brand_mail/models/mail_template.py
